eduwaka_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
cloudinary==1.36.0
colorama==0.4.6
cryptography==46.0.5
dj-database-url==1.3.0
dj-rest-auth==7.2.0
Django==5.2.4
django-allauth==65.15.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
mysql-connector-python==9.3.0
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
